requirements.txt
Flask==2.3.3
PyJWT==2.8.0
Flask-SQLAlchemy==3.1.1
marshmallow==3.21.1
requests==2.31.0
marshmallow-sqlalchemy==0.29.0
flask-cors==6.0.1
Werkzeug==3.1.3
python-dotenv==1.1.0